third_party/blink/web_tests/external/wpt/preload/preload-error.sub-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] 404 (style): preload events
  assert_unreached: preload onerror Reached unreachable code
[FAIL] 404 (script): preload events
  assert_unreached: preload onerror Reached unreachable code
[FAIL] Decode-error (image): preload events
  assert_unreached: preload onerror Reached unreachable code
[FAIL] MIME-error (style): main
  assert_unreached: Should have rejected: undefined Reached unreachable code
[FAIL] MIME-blocked (script): preload events
  assert_unreached: preload onload Reached unreachable code
[FAIL] MIME-blocked-nosniff (script): preload events
  assert_unreached: preload onload Reached unreachable code
Harness: the test ran to completion.
